Move external_id from UserSerializer to CurrentUserSerializer
