+++ app/serializers/user_serializer.rb
@@ -73,8 +73,7 @@ class UserSerializer < BasicUserSerializer
              :primary_group_flair_bg_color,
              :primary_group_flair_color,
              :staged,
-             :second_factor_enabled,
-             :external_id
+             :second_factor_enabled
 
   has_one :invited_by, embed: :object, serializer: BasicUserSerializer
   has_many :groups, embed: :object, serializer: BasicGroupSerializer
@@ -286,14 +285,6 @@ class UserSerializer < BasicUserSerializer
 
   def primary_group_flair_color
     object.try(:primary_group).try(:flair_color)
-  end
-
-  def external_id
-    object&.single_sign_on_record&.external_id
-  end
-
-  def include_external_id?
-    SiteSetting.enable_sso
   end
 
   ###
